Add cpu profiler automator
