+++ packages/devtools_app/benchmark/README.md
@@ -1,8 +1,8 @@
 # DevTools benchmark tests
 
 There are two types of benchmarks that we currently support: size and performance.
-- devtools_benchmarks_test.dart (measures DevTools frame times)
-- web_bundle_size_test.dart (measures DevTools release build size)
+1. `devtools_benchmarks_test.dart` - measures DevTools frame times.
+2. `web_bundle_size_test.dart` - measures DevTools release build size.
 
 The benchmark tests are run automatically on the CI.
 See the "benchmark-performance" and "benchmark-size" jobs.
@@ -14,11 +14,13 @@ See the "benchmark-performance" and "benchmark-size" jobs.
 > (e.g. by running `flutter test benchmark/`). See the [#caveats](#caveats)
 > section below.
 
+All of the commands below should be run from the `packages/devtools_app` directory.
+
 ### Performance benchmarks
 
 To run the performance benchmark tests locally, run:
 ```sh
-dart run run_benchmarks.dart
+dart run benchmark/run_benchmarks.dart
 ```
 
 To run the test that verifies we can run benchmark tests, run:
@@ -39,4 +41,10 @@ The size benchmark must be ran by itself because it actually modifies the
 `devtools_app/build` folder to create and measure the release build web bundle size.
 If this test is ran while other tests are running, it can affect the measurements
 that the size benchmark test takes, and it can affect the DevTools build that
-the other running tests are using with.
+the other running tests are using.
+
+## Adding a new benchmark test or test case
+
+The tests are defined by "automators", which live in the `benchmark/test_infra/automators`
+directory. To add a new test or test case, either modify an existing automator or add
+a new one for a new screen. Follow existing examples in that directory for guidance.
